added debug message to FPGA index read

+++ Control_Subsystem/Control_program/control_program/RoverFunctions.cpp
@@ -31,7 +31,7 @@ Obstacle cartesianToObstacle(RoverDataStructure *rover, ColourObject co, int ind
     double c_x = ((double)co.x - 320)/20;
     double c_y = (480 - (double)co.y)/20*8;
 
-    double angle = atan2((double)c_x, (double)c_y) + ((double)rover->rover_heading*DEG_TO_RAD); //in radians
+    double angle = atan2(c_x, c_y) + ((double)rover->rover_heading*DEG_TO_RAD); //in radians
     double distance = sqrt(sq(c_x) + sq(c_y));
 
     Obstacle tmp;
